CoolMasterNet: Refactor to use Bridge for HVAC controller
