+++ addons/binding/org.openhab.binding.coolmasternet/ESH-INF/binding/binding.xml
@@ -4,8 +4,8 @@
         xmlns:binding="http://eclipse.org/smarthome/schemas/binding/v1.0.0"
         xsi:schemaLocation="http://eclipse.org/smarthome/schemas/binding/v1.0.0 http://eclipse.org/smarthome/schemas/binding-1.0.0.xsd">
 
-    <name>coolmasternet Binding</name>
-    <description>This is the binding for coolmasternet.</description>
+    <name>CoolMasterNet Binding</name>
+    <description>Binding for Cool Automation CoolMasterNet HVAC controllers. Not officially supported by Cool Automation.</description>
     <author>Two Feathers Pty Ltd</author>
 
 </binding:binding>
