+++ addons/binding/org.openhab.binding.coolmasternet/META-INF/MANIFEST.MF
@@ -7,8 +7,7 @@ Bundle-Version: 2.0.0.qualifier
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Bundle-ClassPath: .
 Import-Package: 
- org.openhab.binding.coolmasternet,
- org.openhab.binding.coolmasternet.handler,
+ com.google.common.collect,
  org.eclipse.smarthome.config.core,
  org.eclipse.smarthome.core.library.types,
  org.eclipse.smarthome.core.thing,
@@ -16,6 +15,8 @@ Import-Package:
  org.eclipse.smarthome.core.thing.binding.builder,
  org.eclipse.smarthome.core.thing.type,
  org.eclipse.smarthome.core.types,
+ org.openhab.binding.coolmasternet,
+ org.openhab.binding.coolmasternet.handler,
  org.slf4j
 Service-Component: OSGI-INF/*.xml
 Export-Package: org.openhab.binding.coolmasternet,
